pyproject.toml
[tool.poetry]
name = "ascii_art"
version = "0.1.0"
description = ""
authors = ["Anthony Khoudary <ankhoudary12@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
click = "^7.1.2"
Pillow = "^8.0.1"
numpy = "^1.19.4"

[tool.poetry.dev-dependencies]
flake8 = "^3.8.4"
black = "^20.8b1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
